+++ amplify/backend/api/frontenddeveloperres/parameters.json
@@ -8,5 +8,5 @@
       "Outputs.UserPoolId"
     ]
   },
-  "AuthModeLastUpdated": "2021-11-29T23:30:17.091Z"
+  "AuthModeLastUpdated": "2022-01-13T14:47:01.478Z"
 }
