+++ src/components/page/Login/Login.page.tsx
@@ -1,0 +1,9 @@
+import { Login } from './Login'
+
+export const LoginPage = () => {
+  return (
+    <>
+      <Login />
+    </>
+  )
+}
